# backend/src/schemas/auth.py
from uuid import UUID

from pydantic import BaseModel, Field, field_validator


class LoginRequest(BaseModel):
    email: str
    password: str

    @field_validator("email")
    @classmethod
    def normalize_email(cls, value: str) -> str:
        value = value.strip().lower()
        if "@" not in value:
            msg = "email must contain @"
            raise ValueError(msg)
        return value


class Session(BaseModel):
    user_id: UUID
    role: str
    center_id: UUID | None
    must_change_password: bool


class ChangePasswordRequest(BaseModel):
    current_password: str
    new_password: str = Field(min_length=8)