# specs/001-clinic-scheduling/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Clinic Patient Scheduling API
  version: 0.1.0
  description: >
    Tenant-scoped REST API for clinic patient scheduling. Authentication is via a
    server-side session cookie that carries the user's center scope. Every non-super-admin
    endpoint is scoped to the caller's center and fails closed (403, no data) if scope
    cannot be established (Constitution Principle II). All times are UTC (RFC 3339).
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
  parameters:
...
    doctorId:
      name: doctor_id
      in: query
      schema: { type: string, format: uuid }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: double_booking }
        message: { type: string }
    Role:
      type: string
      enum: [super_admin, center_admin, doctor, reception]
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    Session:
      type: object
      properties:
        user_id: { type: string, format: uuid }
        role: { $ref: '#/components/schemas/Role' }
        center_id: { type: string, format: uuid, nullable: true }
        must_change_password: { type: boolean }
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password: { type: string }
        new_password: { type: string, minLength: 12 }
...
    Appointment:
      type: object
      properties:
        id: { type: string, format: uuid }
        doctor_id: { type: string, format: uuid }
        patient_id: { type: string, format: uuid }
        clinic_id: { type: string, format: uuid, nullable: true }
        starts_at: { type: string, format: date-time }
        duration_minutes: { type: integer }
        status: { type: string, enum: [scheduled, completed, cancelled, no_show] }
        origin: { type: string, enum: [request, direct] }
        source_request_id: { type: string, format: uuid, nullable: true }
    AppointmentCreate:
      type: object
      description: >-
        Provide either patient_id (existing patient) or patient_name +
        patient_clinic_identifier (a walk-in, resolved/created on the fly).
      required: [doctor_id, starts_at, duration_minutes]
      properties:
        doctor_id: { type: string, format: uuid }
        patient_id: { type: string, format: uuid, nullable: true }
        patient_name: { type: string, nullable: true }
        patient_clinic_identifier: { type: string, nullable: true }
        clinic_id: { type: string, format: uuid, nullable: true }
        starts_at: { type: string, format: date-time }
        duration_minutes: { type: integer }
        source_request_id: { type: string, format: uuid, nullable: true }
        confirm_patient_conflict: { type: boolean, default: false }
...
    AppointmentReschedule:
      type: object
      required: [starts_at, duration_minutes]
      properties:
        starts_at: { type: string, format: date-time }
        duration_minutes: { type: integer }
        confirm_patient_conflict: { type: boolean, default: false }
    AppointmentStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [completed, cancelled, no_show] }
        cancel_reason: { type: string, nullable: true }
...
  responses:
    Forbidden:
      description: Not permitted / out of center scope (default-deny)
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: Integrity conflict (e.g., double_booking, slot_taken, invalid_transition)
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unprocessable:
      description: Validation error (e.g., outside_availability, off_grid, patient_conflict)
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
security:
  - sessionCookie: []
paths:
...
  /auth/login:
    post:
      tags: [auth]
      security: []
      summary: Log in; establishes a center-scoped session cookie
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/LoginRequest' } } }
      responses:
        '200': { description: Logged in, content: { application/json: { schema: { $ref: '#/components/schemas/Session' } } } }
        '401': { description: Invalid credentials }
        '403': { description: Account inactive or center suspended }
...
  /auth/session:
    get:
      tags: [auth]
      summary: Current session
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Session' } } } }
        '401': { description: No session }
  /auth/change-password:
    post:
      tags: [auth]
      summary: Change password (also clears must_change_password on first login)
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ChangePasswordRequest' } } }
      responses:
        '204': { description: Changed }
        '422': { $ref: '#/components/responses/Unprocessable' }

  # ---- centers (super-admin only) ----
...
  /appointments:
    get:
      tags: [appointments]
      summary: List appointments (center day view, or per-doctor schedule)
      parameters:
        - { $ref: '#/components/parameters/doctorId' }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
      responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Appointment' } } } } }, '403': { $ref: '#/components/responses/Forbidden' } }
    post:
      tags: [appointments]
      summary: Reception creates a direct (walk-in) booking
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AppointmentCreate' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Appointment' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }      # double_booking / slot_taken
        '422': { $ref: '#/components/responses/Unprocessable' }  # outside_availability / off_grid / patient_conflict
  /appointments/{appointmentId}/reschedule:
    post:
      tags: [appointments]
      summary: Reception reschedules a scheduled appointment (old slot freed atomically)
      parameters: [ { name: appointmentId, in: path, required: true, schema: { type: string, format: uuid } } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AppointmentReschedule' } } } }
      responses: { '200': { description: Rescheduled, content: { application/json: { schema: { $ref: '#/components/schemas/Appointment' } } } }, '409': { $ref: '#/components/responses/Conflict' }, '422': { $ref: '#/components/responses/Unprocessable' } }
  /appointments/{appointmentId}/status:
    put:
      tags: [appointments]
      summary: Reception sets status (completed / cancelled / no_show) per state machine
      parameters: [ { name: appointmentId, in: path, required: true, schema: { type: string, format: uuid } } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AppointmentStatusUpdate' } } } }
      responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Appointment' } } } }, '409': { $ref: '#/components/responses/Conflict' } }

  # ---- notifications ----
...